// frontend/src/types/Registro.ts
export interface Registro {
  id: string;
  primera_interaccion: string;
  ultima_interaccion: string;
  telefono: string;
  nombre: string;
  categoria: string;
  fase: string;
  negocio: string;
  propiedad: string;
  tiempo: string;
  planos: string;
  cita: string;
  calificacion: string;
  resumen: string;
  correo: string;
  seguimiento: string;
}